Split functionality into basic types and I/O.
Use package_config.dart for the basic types.
Use package_config_discovery.dart for the functions which
read, write and find configuration files.
