+++ lib/src/discovery.dart
@@ -5,6 +5,8 @@
 import "dart:io";
 import 'dart:typed_data';
 
+import 'package_config_io.dart';
+
 import "errors.dart";
 import "package_config_impl.dart";
 import "package_config_json.dart";
