+++ test/legacy/all.dart
@@ -2,6 +2,9 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+@deprecated
+library package_config.all_test;
+
 import "package:test/test.dart";
 
 import "discovery_analysis_test.dart" as discovery_analysis;
